refactor: split catalog authoring lifecycle
