+++ apps/backend/src/catalog/authoring/authoringTestSupport.ts
@@ -1,0 +1,62 @@
+import {
+  testAuthorId,
+  testPackageId,
+  testPackageVersionId,
+  testTimestamp,
+} from "../testSupport";
+import type {
+  CatalogPackageRow,
+  CatalogPackageStatus,
+  CatalogPackageVersionRow,
+} from "../types";
+
+export const unsafePublicCatalogStorageReference = `media/blobs/sha256/${"a".repeat(64)}`;
+
+export function createPackageRow(): CatalogPackageRow {
+  return {
+    package_id: testPackageId,
+    author_id: testAuthorId,
+    slug: "spanish-basics",
+    title: "Spanish Basics",
+    summary: "Core Spanish prompts.",
+    description: "Core Spanish flashcards for beginners.",
+    language_tags: ["en", "es"],
+    topic_tags: ["language"],
+    license: "CC-BY-4.0",
+    content_warning: null,
+    cover_package_media_key: null,
+    status: "draft",
+    created_at: testTimestamp,
+    updated_at: testTimestamp,
+    published_at: null,
+    delisted_at: null,
+  };
+}
+
+export function createPackageVersionRow(status: CatalogPackageStatus): CatalogPackageVersionRow {
+  return {
+    package_version_id: testPackageVersionId,
+    package_id: testPackageId,
+    version_number: 1,
+    status,
+    slug: "spanish-basics",
+    title: "Spanish Basics",
+    summary: "Core Spanish prompts.",
+    description: "Core Spanish flashcards for beginners.",
+    language_tags: ["en", "es"],
+    topic_tags: ["language"],
+    license: "CC-BY-4.0",
+    content_warning: null,
+    cover_package_media_key: null,
+    source_workspace_id: null,
+    card_count: 1,
+    created_by_admin_email: "admin@example.com",
+    reviewed_by_admin_email: null,
+    created_at: testTimestamp,
+    updated_at: testTimestamp,
+    submitted_at: null,
+    reviewed_at: null,
+    published_at: null,
+    delisted_at: null,
+  };
+}
